SymSpellChecker.jl

SymSpellChecker.edits!

SymSpellChecker.lookup

SymSpellChecker.transfer_casing_for_matching_text

SymSpellChecker.transfer_casing_for_similar_text

SymSpellChecker.edits! — Method

edits!(delete_words, word, edit_distance, max_dictionary_distance)

Inexpensive and language independent: only deletes, no transposes + replaces + inserts replaces and inserts are expensive and language dependent

SymSpellChecker.lookup — Method

lookup(dict, phrase, max_edit_distance, include_unknown, ignore_token, transfer_casing)

Find suggested spellings for a given phrase word. Parameters ––––– phrase : str The word being spell checked. verbosity : :class:Verbosity The value controlling the quantity/closeness of the returned suggestions. maxeditdistance : int, optional The maximum edit distance between phrase and suggested words. Set to :attr:_max_dictionary_edit_distance by default includeunknown : bool, optional A flag to determine whether to include phrase word in suggestions, if no words within edit distance found. ignoretoken : regex pattern, optional A regex pattern describing what words/phrases to ignore and leave unchanged transfercasing : bool, optional A flag to determine whether the casing –- i.e., uppercase vs lowercase –- should be carried over from phrase. Returns –––- suggestions : list suggestions is a list of :class:SuggestItem objects representing suggested correct spellings for the phrase word, sorted by edit distance, and secondarily by count frequency. Raises ––– ValueError If `maxeditdistanceis greater than :attr:maxdictionaryedit_distance`

SymSpellChecker.transfer_casing_for_matching_text — Method

transfer_casing_for_matching_text(text_w_casing, text_wo_casing)

Transferring the casing from one text to another - assuming that they are 'matching' texts, alias they have the same length. Parameters ––––– textwcasing : str Text with varied casing textwocasing : str Text that is in lowercase only Returns –––- str Text with the content of text_wo_casing and the casing of text_w_casing Raises ––– ArgumentError If the input texts have different lengths

SymSpellChecker.transfer_casing_for_similar_text — Method

transfer_casing_for_similar_text(text_w_casing, text_wo_casing)